\name{Lab4Pro-package}
\alias{Lab4Pro-package}
\alias{Lab4Pro}
\docType{package}
\title{\packageTitle{Lab4Pro}}
\description{\packageDescription{Lab4Pro}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{Lab4Pro}
  \packageIndices{Lab4Pro}

  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
  \packageAuthor{Lab4Pro}
  
  Maintainer: \packageMaintainer{Lab4Pro}
}
\references{
  This optional section can contain literature or other references for
  background information.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
\seealso{
  Optional links to other man pages
}
\examples{
  ## Optional simple examples of the most important functions
  ## Use \dontrun{} around code to be shown but not executed
}
