% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_mean.R
\name{my_mean}
\alias{my_mean}
\title{Mean function}
\usage{
my_mean(x)
}
\arguments{
\item{x}{a list of numeric vectors}
}
\description{
Mean function
}
\details{
Calculates mean of each numeric vector in a list
}
\section{Usecases }{

Count a mean from multiple vectors (or from a data frame columnwise)
}

\examples{
x <- list(c(1,2,3,4,5), rep(1,5), 1:5)
my_mean(x)
}
\seealso{
\code{\link[base]{mean}}
}
\author{
Morgen
}
